+++ package.json
@@ -84,6 +84,7 @@
     "react-dom": "^17.0.1",
     "react-fast-marquee": "^1.3.4",
     "react-image-lightbox": "^5.1.4",
+    "react-intersection-observer": "^9.4.0",
     "react-lazy-load-image-component": "^1.5.4",
     "react-player": "^2.10.0",
     "react-slick": "^0.28.1",
