+++ src/components/accounts/dashboard/MainContent.js
@@ -17,11 +17,8 @@ import ZInstanceItem from './ui/ZInstanceItem';
 import ZMyCard from './ui/ZMyCard';
 import ZTimelineItem from './ui/ZTimelineItem';
 import * as helpers from 'utils';
-import SettingsIcon from '@mui/icons-material/Settings';
-import CodeIcon from '@mui/icons-material/Code';
-import EditIcon from '@mui/icons-material/Edit';
-import StorageIcon from '@mui/icons-material/Storage';
-import BorderColorIcon from '@mui/icons-material/BorderColor';
+
+import ZInstanceTimelineItemContainer from './ui/ZInstanceTimelineItemContainer';
 
 const INSTANCE_LIMIT = 3;
 
@@ -39,120 +36,18 @@ const MainContent = ({
   const [instanceAudit, setInstanceAudit] = useState([]);
   const [isInstanceAuditLoading, setIsInstanceAuditLoading] = useState(false);
 
-  const getIcon = (uri) => {
-    /* if the uri has 
-          items = edit icon
-          settings = cog icon
-          views or scripts or stylesheets = code icon
-          models = schema icon
-      */
-
-    if (uri === undefined) {
-      return <BorderColorIcon color="primary" />;
-    } else if (uri.includes('items')) {
-      return <EditIcon color="primary" />;
-    } else if (uri.includes('settings')) {
-      return <SettingsIcon color="primary" />;
-    } else if (
-      uri.includes('views') ||
-      uri.includes('scripts') ||
-      uri.includes('stylesheets')
-    ) {
-      return <CodeIcon color="primary" />;
-    } else if (uri.includes('models')) {
-      return <StorageIcon color="primary" />;
-    } else {
-      return <BorderColorIcon color="primary" />;
-    }
-  };
-
   const setResponseToAuditState = async (zuid, limit = 10) => {
     const response = await ZestyAPI.getInstanceAuditInitZUID(limit, zuid);
 
     if (!response?.error) {
-      const instanceAuditData = await Promise.all(
-        [...response?.data].map(async (audit) => {
-          let newMessage = '';
-          let auditMessage = audit?.meta?.message;
-          if (
-            auditMessage.includes('Modified Item') &&
-            auditMessage.includes('ContentModel')
-          ) {
-            const itemZUIDName = await ZestyAPI.appendURI(
-              audit?.meta?.uri,
-              audit?.entityZUID,
-            );
-
-            const contentZUIDName = await ZestyAPI.appendURI(
-              audit?.meta?.uri.substr(0, audit.meta.uri.indexOf('/items')),
-              audit?.entityZUID,
-            );
-
-            const message = auditMessage.split('`');
-            const itemZUID = message[1];
-            const contentZUID = message[3];
-            newMessage = auditMessage
-              .replace(itemZUID, itemZUIDName?.data?.web?.metaTitle)
-              .replace(contentZUID, contentZUIDName?.data?.label);
-          } else if (
-            auditMessage.includes('Added Item') &&
-            auditMessage.includes('ContentModel')
-          ) {
-            const message = auditMessage.split('`');
-            const itemZUID = message[1];
-            const contentZUID = message[3];
-
-            const itemZUIDName = await ZestyAPI.appendURI(
-              audit?.meta?.uri.substr(0, audit.meta.uri.indexOf('/items') + 7) +
-                `/${itemZUID}`,
-              audit?.entityZUID,
-            );
-
-            const contentZUIDName = await ZestyAPI.appendURI(
-              audit?.meta?.uri.substr(0, audit.meta.uri.indexOf('/items')),
-              audit?.entityZUID,
-            );
-
-            newMessage = auditMessage
-              .replace(
-                itemZUID,
-                itemZUIDName?.data?.web?.metaTitle ||
-                  itemZUIDName?.data?.data?.title,
-              )
-              .replace(contentZUID, contentZUIDName?.data?.label);
-          } else if (auditMessage.includes('Modified EnvSetting')) {
-            const response = await ZestyAPI.appendURI(
-              audit?.meta?.uri,
-              audit?.entityZUID,
-            );
-
-            const message = auditMessage.split('`');
-            const envZUID = message[1];
-            newMessage = auditMessage.replace(
-              envZUID,
-              response?.data?.keyFriendly,
-            );
-          } else if (auditMessage.includes('Published Item')) {
-            const response = await ZestyAPI.appendURI(
-              audit?.meta?.uri.substr(0, audit.meta.uri.indexOf('/items')),
-              audit?.entityZUID,
-            );
-
-            const message = auditMessage.split('`');
-            const envZUID = message[1];
-            newMessage = auditMessage.replace(envZUID, response?.data?.label);
-          }
-
-          return {
-            ...audit,
-            entityName: instances?.find((i) => i.ZUID === audit.entityZUID)
-              ?.name,
-            screenshotURL: instances?.find((i) => i.ZUID === audit.entityZUID)
-              ?.screenshotURL,
-            newMessage,
-          };
-        }),
-      );
+      const instanceAuditData = [...response?.data].map((audit) => {
+        return {
+          ...audit,
+          entityName: instances?.find((i) => i.ZUID === audit.entityZUID)?.name,
+          screenshotURL: instances?.find((i) => i.ZUID === audit.entityZUID)
+            ?.screenshotURL,
+        };
+      });
 
       setInstanceAudit((prev) =>
         [...prev, ...instanceAuditData].sort(
@@ -257,80 +152,11 @@ const MainContent = ({
                       />
                     ))
                   : instanceAudit?.map((audit, index) => (
-                      <ZTimelineItem
-                        sx={{
-                          '::before': {
-                            content: 'none',
-                          },
-                          mt: 1,
-                        }}
+                      <ZInstanceTimelineItemContainer
                         key={index}
-                        title={`${helpers.getTimeAgo(audit?.updatedAt)}`}
-                        timelineImage={audit?.screenshotURL}
-                        isLoading={isInstanceAuditLoading}
-                      >
-                        <Stack
-                          sx={{ border: `1px solid ${grey[400]}`, p: 2 }}
-                          component={Paper}
-                          elevation={0}
-                          direction={{ xs: 'column', lg: 'row' }}
-                          justifyContent="space-between"
-                        >
-                          <Stack
-                            direction="row"
-                            spacing={2}
-                            alignItems="center"
-                          >
-                            <Stack>{getIcon(audit?.meta?.uri)}</Stack>
-                            <Stack>
-                              <Typography>{`${audit?.entityName} by ${audit?.firstName} ${audit?.lastName}`}</Typography>
-                              <Typography
-                                variant="body2"
-                                sx={{ wordBreak: 'break-word' }}
-                              >
-                                {audit?.newMessage || audit?.meta?.message}
-                              </Typography>
-                            </Stack>
-                          </Stack>
-                          <Stack
-                            alignItems="center"
-                            direction="row"
-                            spacing={2}
-                            mt={{ xs: 1 }}
-                            ml={{ xs: 5, lg: 0 }}
-                          >
-                            {audit?.meta?.url && (
-                              <Button
-                                href={audit?.meta?.url}
-                                size="small"
-                                variant="contained"
-                                color="primary"
-                                sx={{
-                                  whiteSpace: {
-                                    md: 'nowrap',
-                                  },
-                                }}
-                              >
-                                Open Resource
-                              </Button>
-                            )}
-
-                            <Button
-                              href={`/instances/${audit?.entityZUID}`}
-                              size="small"
-                              variant="outlined"
-                              color="primary"
-                              sx={{
-                                whiteSpace: {
-                                  md: 'nowrap',
-                                },
-                              }}
-                            >
-                              Edit Instance
-                            </Button>
-                          </Stack>
-                        </Stack>
-                      </ZTimelineItem>
+                        audit={audit}
+                        isInstanceAuditLoading={isInstanceAuditLoading}
+                      />
                     ))}
               </Timeline>
             )}
